+++ src/Nest/Domain/Stats/GetStats.cs
@@ -19,14 +19,14 @@ namespace Nest
 		public long Current { get; set; }
 
 		[JsonProperty(PropertyName = "exists_total")]
-		public int ExistsTotal { get; set; }
+		public long ExistsTotal { get; set; }
 		[JsonProperty(PropertyName = "exists_time")]
 		public string ExistsTime { get; set; }
 		[JsonProperty(PropertyName = "exists_time_in_millis")]
 		public double ExistsTimeInMilliseconds { get; set; }
 
 		[JsonProperty(PropertyName = "missing_total")]
-		public int MissingTotal { get; set; }
+		public long MissingTotal { get; set; }
 		[JsonProperty(PropertyName = "missing_time")]
 		public string MissingTime { get; set; }
 		[JsonProperty(PropertyName = "missing_time_in_millis")]
